Deleted team and ball from evaluation directory and refactored some method calls
