+++ src/software/ai/hl/stp/action/BUILD
@@ -7,12 +7,12 @@ cc_library(
     deps = [
         ":action",
         "//shared:constants",
-        "//software/ai/evaluation:ball",
         "//software/ai/intent:chip_intent",
         "//software/ai/intent:move_intent",
         "//software/geom:polygon",
         "//software/geom:triangle",
         "//software/geom/algorithms",
+        "//software/world:ball",
     ],
 )
 
@@ -64,7 +64,6 @@ cc_library(
     deps = [
         ":action",
         "//shared:constants",
-        "//software/ai/evaluation:ball",
         "//software/ai/intent:kick_intent",
         "//software/ai/intent:move_intent",
         "//software/geom:polygon",
