+++ src/software/ai/hl/stp/action/chip_action.cpp
@@ -1,12 +1,12 @@
 #include "software/ai/hl/stp/action/chip_action.h"
 
 #include "shared/constants.h"
-#include "software/ai/evaluation/ball.h"
 #include "software/ai/intent/chip_intent.h"
 #include "software/ai/intent/move_intent.h"
 #include "software/geom/algorithms/contains.h"
 #include "software/geom/polygon.h"
 #include "software/geom/triangle.h"
+#include "software/world/ball.h"
 
 ChipAction::ChipAction() : Action(false), ball({0, 0}, {0, 0}, Timestamp::fromSeconds(0))
 {
@@ -130,5 +130,5 @@ void ChipAction::calculateNextIntent(IntentCoroutine::push_type& yield)
             yield(std::make_unique<ChipIntent>(robot->id(), chip_origin, chip_direction,
                                                chip_distance_meters, 0));
         }
-    } while (!hasBallBeenKicked(ball, chip_direction));
+    } while (!ball.hasBallBeenKicked(chip_direction));
 }
