+++ src/software/ai/hl/stp/action/kick_action.cpp
@@ -1,7 +1,6 @@
 #include "software/ai/hl/stp/action/kick_action.h"
 
 #include "shared/constants.h"
-#include "software/ai/evaluation/ball.h"
 #include "software/ai/intent/kick_intent.h"
 #include "software/ai/intent/move_intent.h"
 #include "software/geom/algorithms/contains.h"
@@ -127,5 +126,5 @@ void KickAction::calculateNextIntent(IntentCoroutine::push_type &yield)
             yield(std::make_unique<KickIntent>(robot->id(), kick_origin, kick_direction,
                                                kick_speed_meters_per_second, 0));
         }
-    } while (!hasBallBeenKicked(ball, kick_direction));
+    } while (!ball.hasBallBeenKicked(kick_direction));
 }
