+++ src/software/ai/hl/stp/play/corner_kick_play.cpp
@@ -1,7 +1,6 @@
 #include "software/ai/hl/stp/play/corner_kick_play.h"
 
 #include "shared/constants.h"
-#include "software/ai/evaluation/ball.h"
 #include "software/ai/evaluation/possession.h"
 #include "software/ai/hl/stp/tactic/goalie_tactic.h"
 #include "software/ai/hl/stp/tactic/move_tactic.h"
@@ -10,6 +9,7 @@
 #include "software/ai/passing/pass_generator.h"
 #include "software/logger/logger.h"
 #include "software/util/design_patterns/generic_factory.h"
+#include "software/world/ball.h"
 
 CornerKickPlay::CornerKickPlay()
     : MAX_TIME_TO_COMMIT_TO_PASS(Duration::fromSeconds(DynamicParameters->getAIConfig()
