+++ src/software/ai/hl/stp/tactic/BUILD
@@ -191,12 +191,12 @@ cc_library(
     deps = [
         ":tactic",
         "//shared:constants",
-        "//software/ai/evaluation:ball",
         "//software/ai/hl/stp/action:intercept_ball_action",
         "//software/ai/hl/stp/action:kick_action",
         "//software/ai/hl/stp/action:move_action",
         "//software/ai/passing:pass",
         "//software/logger",
+        "//software/world:ball",
     ],
 )
