+++ src/software/world/team.cpp
@@ -182,13 +182,13 @@ std::vector<Robot> Team::getAllRobotsExceptGoalie() const
 ////////////////////
 
 
-std::optional<Robot> Team::getNearestRobot(const Point &ref_point) const
+std::optional<Robot> Team::getNearestRobot(const Point& ref_point) const
 {
     return getNearestRobot(this->getAllRobots(), ref_point);
 }
 
-std::optional<Robot> Team::getNearestRobot(const std::vector<Robot> &robots,
-                                           const Point &ref_point) const
+std::optional<Robot> Team::getNearestRobot(const std::vector<Robot>& robots,
+                                           const Point& ref_point) const
 {
     if (robots.empty())
     {
@@ -196,7 +196,7 @@ std::optional<Robot> Team::getNearestRobot(const std::vector<Robot> &robots,
     }
 
     Robot nearest_robot = robots.at(0);
-    for (const Robot &curRobot : robots)
+    for (const Robot& curRobot : robots)
     {
         double curDistance = (ref_point - curRobot.position()).length();
         if (curDistance < (nearest_robot.position() - ref_point).length())
