+++ src/software/world/team.h
@@ -171,7 +171,7 @@ class Team
      * @param ref_point The point where the distance to each robot will be measured.
      * @return Robot that is closest to the reference point.
      */
-    std::optional<Robot> getNearestRobot(const Point &ref_point) const;
+    std::optional<Robot> getNearestRobot(const Point& ref_point) const;
 
     /**
      * Given a list of robots, finds the robot on that team that is closest to a
@@ -181,8 +181,8 @@ class Team
      * @param ref_point The point where the distance to each robot will be measured.
      * @return Robot that is closest to the reference point.
      */
-    std::optional<Robot> getNearestRobot(const std::vector<Robot> &robots,
-                                         const Point &ref_point) const;
+    std::optional<Robot> getNearestRobot(const std::vector<Robot>& robots,
+                                         const Point& ref_point) const;
 
 
     /////////////////////
